- Change Struct Location
